CRUD work -restotan/vote/user vote LAZY bad

+++ src/main/java/ru/javawebinar/vote/service/RestoranService.java
@@ -21,7 +21,7 @@ public class RestoranService {
         return repository.findAll();
     }
 
-    public Restoran get(Long id) {
+    public Restoran get(int id) {
         return repository.getById(id);
     }
